fix(admin-ui): implement clean code architecture in schema pluign
