+++ admin-ui/plugins/schema/domain/entities/TDefautAttribute.js
@@ -1,0 +1,15 @@
+// export interface TDefautAttribute {
+//   jansHideOnDiscovery: boolean
+//   selected: boolean
+//   scimCustomAttr: boolean
+//   oxMultiValuedAttribute: boolean
+//   custom: boolean
+//   requred: boolean
+//   attributeValidation: TAttributeValidation
+// }
+
+// interface TAttributeValidation {
+//   maxLength: any
+//   regexp: any
+//   minLength: any
+// }
