+++ admin-ui/plugins/schema/plugin-metadata.js
@@ -1,9 +1,9 @@
-import AttributeListPage from './components/Person/AttributeListPage'
-import AttributeAddPage from './components/Person/AttributeAddPage'
-import AttributeEditPage from './components/Person/AttributeEditPage'
-import AttributeViewPage from './components/Person/AttributeViewPage'
-import { reducer as attributeReducer } from './redux/features/attributeSlice'
-import attributeSaga from './redux/sagas/AttributeSaga'
+import AttributeListPage from './presentation/pages/AttributeListPage'
+import AttributeAddPage from './presentation/pages/AttributeAddPage'
+import AttributeEditPage from './presentation/pages/AttributeEditPage'
+import AttributeViewPage from './presentation/pages/AttributeViewPage'
+import { reducer as attributeReducer } from './infrastructure/redux/features/attributeSlice'
+import attributeSaga from './infrastructure/redux/sagas/AttributeSaga'
 import { ATTRIBUTE_READ, ATTRIBUTE_WRITE } from 'Utils//PermChecker'
 
 const pluginMetadata = {
