+++ admin-ui/plugins/schema/presentation/components/Person/AttributeForm.js
@@ -1,0 +1,373 @@
+import React, { useState } from 'react'
+import { Formik, ErrorMessage } from 'formik'
+import * as Yup from 'yup'
+import {
+  Col,
+  InputGroup,
+  CustomInput,
+  Form,
+  FormGroup,
+  Input
+} from 'Components'
+import GluuFooter from 'Routes/Apps/Gluu/GluuFooter'
+import GluuLabel from 'Routes/Apps/Gluu/GluuLabel'
+import GluuInumInput from 'Routes/Apps/Gluu/GluuInumInput'
+import GluuInputRow from 'Routes/Apps/Gluu/GluuInputRow'
+import GluuToogleRow from 'Routes/Apps/Gluu/GluuToogleRow'
+import { ATTRIBUTE } from 'Utils/ApiResources'
+import { useTranslation } from 'react-i18next'
+import {
+  getInitialState,
+  handleAttributeSubmit,
+  getInitialAttributeValues,
+  attributeValidationSchema
+} from 'Plugins/schema/domain/use-cases/attributeUseCases'
+
+function AttributeForm({ item, customOnSubmit, hideButtons }) {
+  const { t } = useTranslation()
+  const [init, setInit] = useState(false)
+  const [validation, setValidation] = useState(getInitialState(item))
+  function handleValidation() {
+    setValidation(!validation)
+  }
+  function toogle() {
+    if (!init) {
+      setInit(true)
+    }
+  }
+
+  return (
+    <Formik
+      initialValues={getInitialAttributeValues(item)}
+      validationSchema={attributeValidationSchema}
+      onSubmit={(values) =>
+        handleAttributeSubmit({ values, item, customOnSubmit })
+      }
+    >
+      {(formik) => (
+        <Form onSubmit={formik.handleSubmit}>
+          {item.inum && (
+            <GluuInumInput
+              label='fields.inum'
+              name='inum'
+              lsize={3}
+              rsize={9}
+              value={item.inum}
+              doc_category={ATTRIBUTE}
+            />
+          )}
+          <FormGroup row>
+            <GluuLabel
+              label='fields.name'
+              required
+              doc_category={ATTRIBUTE}
+              doc_entry='name'
+            />
+            <Col sm={9}>
+              <Input
+                placeholder={t('placeholders.enter_the_attribute_name')}
+                id='name'
+                valid={!formik.errors.name && !formik.touched.name && init}
+                name='name'
+                defaultValue={item.name}
+                onKeyUp={toogle}
+                onChange={formik.handleChange}
+              />
+              <ErrorMessage name='name'>
+                {(msg) => <div style={{ color: 'red' }}>{msg}</div>}
+              </ErrorMessage>
+            </Col>
+          </FormGroup>
+
+          <FormGroup row>
+            <GluuLabel
+              label='fields.displayname'
+              required
+              doc_category={ATTRIBUTE}
+              doc_entry='displayName'
+            />
+            <Col sm={9}>
+              <InputGroup>
+                <Input
+                  placeholder={t(
+                    'placeholders.enter_the_attribute_display_name'
+                  )}
+                  valid={
+                    !formik.errors.displayName &&
+                    !formik.touched.displayName &&
+                    init
+                  }
+                  id='displayName'
+                  name='displayName'
+                  defaultValue={item.displayName}
+                  onChange={formik.handleChange}
+                />
+              </InputGroup>
+              <ErrorMessage name='displayName'>
+                {(msg) => <div style={{ color: 'red' }}>{msg}</div>}
+              </ErrorMessage>
+            </Col>
+          </FormGroup>
+
+          <FormGroup row>
+            <GluuLabel
+              label='fields.description'
+              required
+              doc_category={ATTRIBUTE}
+              doc_entry='description'
+            />
+            <Col sm={9}>
+              <InputGroup>
+                <Input
+                  type='textarea'
+                  rows='3'
+                  placeholder={t(
+                    'placeholders.enter_the_attribute_description'
+                  )}
+                  id='description'
+                  name='description'
+                  defaultValue={item.description}
+                  onChange={formik.handleChange}
+                />
+              </InputGroup>
+              <ErrorMessage name='description'>
+                {(msg) => <div style={{ color: 'red' }}>{msg}</div>}
+              </ErrorMessage>
+            </Col>
+          </FormGroup>
+
+          <FormGroup row>
+            <GluuLabel
+              label='fields.status'
+              required
+              doc_category={ATTRIBUTE}
+              doc_entry='status'
+            />
+            <Col sm={9}>
+              <InputGroup>
+                <CustomInput
+                  type='select'
+                  id='status'
+                  name='status'
+                  defaultValue={item.status}
+                  onChange={formik.handleChange}
+                >
+                  <option value=''>{t('options.choose')}...</option>
+                  <option value='active'>{t('options.active')}</option>
+                  <option value='inactive'>{t('options.inactive')}</option>
+                </CustomInput>
+              </InputGroup>
+              <ErrorMessage name='status'>
+                {(msg) => <div style={{ color: 'red' }}>{msg}</div>}
+              </ErrorMessage>
+            </Col>
+          </FormGroup>
+
+          <FormGroup row>
+            <GluuLabel
+              label='fields.data_type'
+              required
+              doc_category={ATTRIBUTE}
+              doc_entry='dataType'
+            />
+            <Col sm={9}>
+              <InputGroup>
+                <CustomInput
+                  type='select'
+                  id='dataType'
+                  name='dataType'
+                  defaultValue={item.dataType}
+                  onChange={formik.handleChange}
+                >
+                  <option value=''>{t('options.choose')}...</option>
+                  <option value='string'>{t('options.string')}</option>
+                  <option value='json'>{t('options.json')}</option>
+                  <option value='numeric'>{t('options.numeric')}</option>
+                  <option value='binary'>{t('options.binary')}</option>
+                  <option value='certificate'>
+                    {t('options.certificate')}
+                  </option>
+                  <option value='date'>{t('options.date')}</option>
+                  <option value='boolean'>{t('options.boolean')}</option>
+                </CustomInput>
+              </InputGroup>
+              <ErrorMessage name='dataType'>
+                {(msg) => <div style={{ color: 'red' }}>{msg}</div>}
+              </ErrorMessage>
+            </Col>
+          </FormGroup>
+
+          <FormGroup row>
+            <GluuLabel
+              label='fields.edit_type'
+              required
+              doc_category={ATTRIBUTE}
+              doc_entry='editType'
+            />
+            <Col sm={9}>
+              <Input
+                type='select'
+                name='editType'
+                id='editType'
+                defaultValue={item.editType}
+                multiple
+                onChange={formik.handleChange}
+              >
+                <option value='admin'>{t('options.admin')}</option>
+                <option value='user'>{t('options.user')}</option>
+              </Input>
+              <ErrorMessage name='editType'>
+                {(msg) => <div style={{ color: 'red' }}>{msg}</div>}
+              </ErrorMessage>
+            </Col>
+          </FormGroup>
+
+          <FormGroup row>
+            <GluuLabel
+              label='fields.view_type'
+              doc_category={ATTRIBUTE}
+              doc_entry='viewType'
+            />
+            <Col sm={9}>
+              <Input
+                type='select'
+                name='viewType'
+                id='viewType'
+                defaultValue={item.viewType}
+                multiple
+                onChange={formik.handleChange}
+              >
+                <option value='admin'>{t('options.admin')}</option>
+                <option value='user'>{t('options.user')}</option>
+              </Input>
+            </Col>
+          </FormGroup>
+
+          <FormGroup row>
+            <GluuLabel
+              label='fields.usage_type'
+              required
+              doc_category={ATTRIBUTE}
+              doc_entry='usageType'
+            />
+            <Col sm={9}>
+              <Input
+                type='select'
+                name='usageType'
+                id='usageType'
+                defaultValue={item.usageType}
+                multiple
+                onChange={formik.handleChange}
+              >
+                <option value='openid'>{t('options.openid')}</option>
+              </Input>
+              <ErrorMessage name='usageType'>
+                {(msg) => <div style={{ color: 'red' }}>{msg}</div>}
+              </ErrorMessage>
+            </Col>
+          </FormGroup>
+          <GluuInputRow
+            label='fields.oxauth_claim_name'
+            name='claimName'
+            formik={formik}
+            value={item.claimName}
+            doc_category={ATTRIBUTE}
+          />
+          <FormGroup row>
+            <Col sm={4}>
+              <GluuToogleRow
+                name='oxMultiValuedAttribute'
+                formik={formik}
+                lsize={6}
+                rsize={6}
+                label='fields.multivalued'
+                value={item.oxMultiValuedAttribute}
+                doc_category={ATTRIBUTE}
+              />
+            </Col>
+            <Col sm={4}>
+              <GluuToogleRow
+                name='jansHideOnDiscovery'
+                formik={formik}
+                lsize={6}
+                rsize={6}
+                label='fields.hide_on_discovery'
+                value={item.jansHideOnDiscovery}
+                doc_category={ATTRIBUTE}
+              />
+            </Col>
+            <Col sm={4}>
+              <GluuToogleRow
+                name='scimCustomAttr'
+                formik={formik}
+                lsize={6}
+                rsize={6}
+                label='fields.include_in_scim_extension'
+                value={item.scimCustomAttr}
+                doc_category={ATTRIBUTE}
+              />
+            </Col>
+          </FormGroup>
+          <GluuToogleRow
+            name='validation'
+            formik={formik}
+            label='fields.enable_custom_validation_for_this_attribute'
+            value={validation}
+            handler={handleValidation}
+            doc_category={ATTRIBUTE}
+          />
+          {validation && (
+            <GluuInputRow
+              label='fields.regular_expression'
+              name='regexp'
+              formik={formik}
+              value={item.attributeValidation.regexp}
+              doc_category={ATTRIBUTE}
+            />
+          )}
+          {validation && (
+            <FormGroup row>
+              <Col sm={6}>
+                <GluuInputRow
+                  label='fields.minimum_length'
+                  name='minLength'
+                  formik={formik}
+                  type='number'
+                  value={item.attributeValidation.minLength}
+                  doc_category={ATTRIBUTE}
+                />
+              </Col>
+              <Col sm={6}>
+                <GluuInputRow
+                  label='fields.maximum_length'
+                  name='maxLength'
+                  formik={formik}
+                  type='number'
+                  value={item.attributeValidation.maxLength}
+                  doc_category={ATTRIBUTE}
+                />
+              </Col>
+            </FormGroup>
+          )}
+          <GluuInputRow
+            label='fields.saml1_uri'
+            name='saml1Uri'
+            formik={formik}
+            value={item.saml1Uri}
+            doc_category={ATTRIBUTE}
+          />
+          <GluuInputRow
+            label='fields.saml2_uri'
+            name='saml2Uri'
+            formik={formik}
+            value={item.saml2Uri}
+            doc_category={ATTRIBUTE}
+          />
+          <GluuFooter hideButtons={hideButtons} />
+        </Form>
+      )}
+    </Formik>
+  )
+}
+
+export default AttributeForm
