+++ admin-ui/plugins/schema/presentation/pages/AttributeAddPage.js
@@ -2,9 +2,10 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { useNavigate } from 'react-router-dom'
 import { CardBody, Card } from 'Components'
-import AttributeForm from './AttributeForm'
-import { addAttribute } from 'Plugins/schema/redux/features/attributeSlice'
+import AttributeForm from '../components/Person/AttributeForm'
+import { addAttribute } from 'Plugins/schema/infrastructure/redux/features/attributeSlice'
 import applicationStyle from 'Routes/Apps/Gluu/styles/applicationstyle'
+import { defautAttribute } from 'Plugins/schema/domain/use-cases/attributeUseCases'
 
 function AttributeAddPage({ dispatch }) {
   const navigate =useNavigate()
@@ -14,15 +15,7 @@ function AttributeAddPage({ dispatch }) {
       navigate('/attributes')
     }
   }
-  const defautAttribute = {
-    jansHideOnDiscovery: false,
-    selected: false,
-    scimCustomAttr: false,
-    oxMultiValuedAttribute: false,
-    custom: false,
-    requred: false,
-    attributeValidation: { maxLength: null, regexp: null, minLength: null },
-  }
+  
   return (
     <React.Fragment>
       <Card className="mb-3" style={applicationStyle.mainCard}>
