+++ admin-ui/plugins/schema/presentation/pages/AttributeDetailPage.js
